% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE29721}
\alias{GSE29721}
\title{The landscape of promoter DNA hypomethylation in liver cancer
(expression data}
\format{
A data frame with columns:
\describe{
\item{Gene symbol}{HGNC gene symbols}
\item{GSM737065}{10_cancer}
\item{GSM737066}{10_normal}
\item{GSM737067}{11_cancer}
\item{GSM737068}{11_normal}
\item{GSM737069}{12_cancer}
\item{GSM737070}{12_normal}
\item{GSM737071}{14_cancer}
\item{GSM737072}{14_normal}
\item{GSM737073}{15_cancer}
\item{GSM737074}{15_normal}
\item{GSM737075}{1_cancer}
\item{GSM737076}{1_normal}
\item{GSM737077}{4_cancer}
\item{GSM737078}{4_normal}
\item{GSM737079}{5_cancer}
\item{GSM737080}{5_normal}
\item{GSM737081}{8_cancer}
\item{GSM737082}{8_normal}
\item{GSM737083}{9_cancer}
\item{GSM737084}{9_normal}
}
}
\source{
McGill University, Canada.
}
\usage{
GSE29721
}
\description{
Cancerous and normal adjacent tissue samples were obtained from 11 patients
with HCC from the Chinese National Human Genome Center at Shanghai,
China (Dr. Ze-Guang Han). For three patients, the cancer samples were
dissected using the laser capture microdissection technique. Gene
expression profiles for these patients were generated using Affymetrix
expression microarrays. The data is in rda format.
}
\examples{
\dontrun{
 GSE29721
}
}
\references{
Stefanska B, Huang J, Bhattacharyya B, Suderman M et al. Definition of the
landscape of promoter DNA hypomethylation in liver cancer. Cancer Res
2011 Sep 1;71(17):5891-903. PMID: 21747116
}
\keyword{datasets}
