% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE84402}
\alias{GSE84402}
\title{Expression data from human hepatocellular carcinoma tissues and the
correponding non-cancerous tissues}
\format{
A data frame with columns:
\describe{
\item{Gene symbol}{HGNC gene symbols}
\item{GSM2233086}{sample1,cancertissue,male}
\item{GSM2233087}{sample2,non-canceroustissue,male}
\item{GSM2233088}{sample3,cancertissue,male}
\item{GSM2233089}{sample4,non-canceroustissue,male}
\item{GSM2233090}{sample5,cancertissue,male}
\item{GSM2233091}{sample6,non-canceroustissue,male}
\item{GSM2233092}{sample11,cancertissue,male}
\item{GSM2233093}{sample12,non-canceroustissue,male}
\item{GSM2233094}{sample15,cancertissue,male}
\item{GSM2233095}{sample16,non-canceroustissue,male}
\item{GSM2233096}{sample17,cancertissue,male}
\item{GSM2233097}{sample18,non-canceroustissue,male}
\item{GSM2233098}{sample21,cancertissue,female}
\item{GSM2233099}{sample22,non-canceroustissue,female}
\item{GSM2233100}{sample23,cancertissue,female}
\item{GSM2233101}{sample24,non-canceroustissue,female}
\item{GSM2233102}{sample25,cancertissue,female}
\item{GSM2233103}{sample26,non-canceroustissue,female}
\item{GSM2233104}{sample27,cancertissue,female}
\item{GSM2233105}{sample28,non-canceroustissue,female}
\item{GSM2233106}{sample29,cancertissue,female}
\item{GSM2233107}{sample30,non-canceroustissue,female}
\item{GSM2233108}{sample31,cancertissue,male}
\item{GSM2233109}{sample32,non-canceroustissue,male}
\item{GSM2233110}{sample33,cancertissue,male}
\item{GSM2233111}{sample34,non-canceroustissue,male}
\item{GSM2233112}{sample35,cancertissue,male}
\item{GSM2233113}{sample36,non-canceroustissue,male}
}
}
\source{
Shanghai Cancer Insitute, China.
}
\usage{
GSE84402
}
\description{
Total RNA from 14 pairs hepatocellular carcinoma tissues and correponding
non-cancerous tissues were isolated and purified. Five micrograms of
total RNA from each sample were used to synthesize double-stranded
complementary DNA (cDNA), and then amplified and labeled with biotin
and hybridized to the GeneChip Human Genome U133 Plus 2.0 Array
}
\examples{
\dontrun{
 GSE84402
}
}
\references{
Wang H, Huo X, Yang XR, He J et al. STAT3-mediated upregulation of lncRNA
HOXD-AS1 as a ceRNA facilitates liver cancer metastasis by regulating
SOX4. Mol Cancer 2017 Aug 14;16(1):136. PMID: 28810927
}
\keyword{datasets}
