% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logFCsingle.R
\name{logFCsingle}
\alias{logFCsingle}
\title{number of column in expressionLevel should equal to length of setUp}
\usage{
logFCsingle(expressionLevel, setUp)
}
\description{
number of column in expressionLevel should equal to length of setUp
}
