% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatMap.R
\name{plotHeatMap}
\alias{plotHeatMap}
\title{drawing heatmap of gene differential expression data}
\usage{
plotHeatMap(data, nUp = -1, nDown = -1)
}
\arguments{
\item{data}{A dataframe where each row is a gene(row name is the gene
symbol),and each column is the logFC of different
studies(column name is the study name). The data frame must be in
order, where the gene with highest logFC value acorss all studies
should be at the top, and the gene with lowest logFC value acorss all
studies should be at the bottom}

\item{nUp}{A numeric vector indicate the top n numbers of genes with
significant logFC values from up-regulated data. If both nUp and ndown
is not given, all of the data will be plot. Notice that n must be
smaller or equal the number of genes that are up-regulated in data.}

\item{ndown}{A numeric vector indicate the top n numbers of genes with
significant logFC values from down-regulated data. If both nUp and ndown
is not given, all of the data will be plot. Notice that n must be
smaller or equal the number of genes that are down-regulated in data.}
}
\value{
This function will return 1 if the function has been executed
successfully, and return -1 if error occurred
}
\description{
plotHeatMap is a function used to draw a heatmap representation of
differential gene expression.
}
\details{
This function has no return value, the after the function being executed,
a heatmap file named "logFC.tiff" will be saved into current directory.
}
\examples{
\dontrun{
# compute logFC for data1
class <- c("mutant", "control","mutant", "control","mutant", "control",
    "mutant", "control","mutant", "control","mutant", "control", "mutant",
    "control","mutant", "control","mutant", "control", "mutant", "control")
result <- logFCsingle(GSE29721, class)

# compute logFC for data2
class <- c("mutant", "control","mutant", "control","mutant", "control",
    "mutant", "control","mutant", "control","mutant", "control", "mutant",
    "control","mutant", "control","mutant", "control", "mutant", "control",
    "mutant", "control","mutant", "control", "mutant", "control","mutant",
    "control")
result2 <- logFCsingle(GSE84402, class)

# analysis
listLogFC <- list(result, result2)
listTitle <- c("GSE29721", "GSE84402")
aggreg <- Aggreg(listLogFC, listTitle, padj = 0.01, logFC = 1)

# drawing heatmap
plotHeatMap(data.frame(aggreg[3]))

}

}
\references{
Kolde R (2019). \emph{pheatmap: Pretty Heatmaps}.
R package version 1.0.12, \url{https://CRAN.R-project.org/package=pheatmap}.

Wickham H, François R, Henry L, Müller K (2022).
\emph{dplyr: A Grammar of Data Manipulation}.
R package version 1.0.10, \url{https://CRAN.R-project.org/package=dplyr}.

Wang H, Huo X, Yang XR, He J et al. STAT3-mediated upregulation of lncRNA
HOXD-AS1 as a ceRNA facilitates liver cancer metastasis by regulating
SOX4. Mol Cancer 2017 Aug 14;16(1):136. PMID: 28810927

Stefanska B, Huang J, Bhattacharyya B, Suderman M et al. Definition of the
landscape of promoter DNA hypomethylation in liver cancer. Cancer Res
2011 Sep 1;71(17):5891-903. PMID: 21747116
}
