% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDeregGenes.R
\name{runDeregGenes}
\alias{runDeregGenes}
\title{Launch Shiny App for DeregGenes}
\usage{
runDeregGenes()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for DeregGenes.
The shiny app permit to plot heatmap, and visualized analysis result of the
package.
}
\examples{
\dontrun{
runDeregGenes()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
