% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggreg.R
\name{Aggreg}
\alias{Aggreg}
\title{aggregate gene differential expression from different studies together}
\usage{
Aggreg(listLogFC, listTitle, padj = 0.01, logFC = 1)
}
\arguments{
\item{listLogFC}{A vector of all the studies gene differential expression
data. For each study's data, the format should be in data frame, and
each row represent different genes(with the gene symbol as its row
name). The first should be the expressional change(logFC)}

\item{listTitle}{A vector of strings. The length of listTile should be the
same of the length of the listLogFC. listTile should consist of the
title/name of the studies as in the same orders of the studies in
listLogFC}

\item{padj}{A numeric vector that represent the adjust p value threshold}

\item{logFC}{A numeric vector that represent the logFC threshold}
}
\value{
Return a list of length three. The first element of the list is
a data frame of up-regulated differential genes. The second element of
the list is a data frame of down-regulated differential genes. For the
first two data frame, each row a a different genes, and there will be
four columns: gene symbol(Name), p value(Pvalue), adjust p
value(adjPvalue), and expressional change(logFC). The last element of
the list is a aggregated data frame where each row is a gene,
and each column is the logFC of different studies.
}
\description{
Aggreg is a function to aggregate different gene expression fold changes
across different studies. This function will only keep the genes that
all the studies have data of.
}
\examples{
# Require download of about 300MB file.
\dontrun{
# download data1 from GEO
filePaths <- getGEOSuppFiles("GSE29721")

# untar downloaded data1 and delete tar file
untarPath <- strsplit(row.names(filePaths), '/')
untarPath <- paste(untarPath[[1]][1:length(untarPath[[1]]) - 1],
    collapse="/")
untar(row.names(filePaths), exdir = untarPath)
unlink(paste(untarPath, '/*.tar', sep = ''))

# preparing data1
data1 <- prepareData(untarPath, TRUE)

# compute logFC for data1
class <- c("mutant", "control","mutant", "control","mutant", "control",
    "mutant", "control","mutant", "control","mutant", "control", "mutant",
    "control","mutant", "control","mutant", "control", "mutant", "control")
result <- logFCsingle(data1, class)

# delete all download data1
unlink(untarPath, recursive = TRUE)

# download data2 from GEO
filePaths <- getGEOSuppFiles("GSE84402")

# untar downloaded data2 and delete tar file
untarPath <- strsplit(row.names(filePaths), '/')
untarPath <- paste(untarPath[[1]][1:length(untarPath[[1]]) - 1],
    collapse="/")
untar(row.names(filePaths), exdir = untarPath)
unlink(paste(untarPath, '/*.tar', sep = ''))

# preparing data2
data2 <- prepareData(untarPath, TRUE)

# compute logFC for data2
class <- c("mutant", "control","mutant", "control","mutant", "control",
    "mutant", "control","mutant", "control","mutant", "control", "mutant",
    "control","mutant", "control","mutant", "control", "mutant", "control",
    "mutant", "control","mutant", "control", "mutant", "control","mutant",
    "control")
result2 <- logFCsingle(data2, class)

# delete all download data2
unlink(untarPath, recursive = TRUE)


# analysis
listLogFC <- list(result, result2)
listTitle <- c("GSE29721", "GSE84402")
aggreg <- Aggreg(listLogFC, listTitle)

}

}
\references{
Kolde R (2022). \emph{RobustRankAggreg: Methods for Robust Rank Aggregation}.
R package version 1.2.1,
\url{https://CRAN.R-project.org/package=RobustRankAggreg}.
}
