% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggreg.R
\name{Aggreg}
\alias{Aggreg}
\title{Aggregate gene differential expression from different studies together}
\usage{
Aggreg(listLogFC, listTitle, padj = 0.01, logFC = 1)
}
\arguments{
\item{listLogFC}{A vector of all the studies gene differential expression
data. For each study's data, the format should be in data frame, and
each row represent different genes(with the gene symbol as its row
name). The first should be the expressional change(logFC). The length of
this list should be at least 2 for the purpose of this function}

\item{listTitle}{A vector of strings. The length of listTile should be the
same of the length of the listLogFC. listTile should consist of the
title/name of the studies as in the same orders of the studies in
listLogFC}

\item{padj}{A numeric vector that represent the adjust p value threshold,
if nothing is entered, the default value of 0.01 will be used.}

\item{logFC}{A numeric vector that represent the logFC threshold, if
nothing is entered, the default value of 1 will be used.}
}
\value{
Return a list of length three. The first element of the list is
a data frame of up-regulated differential genes. The second element of
the list is a data frame of down-regulated differential genes. For the
first two data frame, each row a a different genes, and there will be
four columns: gene symbol(Name), p value(Pvalue), adjust p
value(adjPvalue), and expressional change(logFC). The last element of
the list is a aggregated data frame where each row is a gene,
and each column is the logFC of different studies.
}
\description{
Aggreg is a function to aggregate different gene expression fold changes
across different studies. This function will only keep the genes that
all the studies have data of.
}
\examples{
\dontrun{
# compute logFC for data1
class <- c("mutant", "control","mutant", "control","mutant", "control",
    "mutant", "control","mutant", "control","mutant", "control", "mutant",
    "control","mutant", "control","mutant", "control", "mutant", "control")
result <- logFCsingle(GSE29721, class)

# compute logFC for data2
class <- c("mutant", "control","mutant", "control","mutant", "control",
    "mutant", "control","mutant", "control","mutant", "control", "mutant",
    "control","mutant", "control","mutant", "control", "mutant", "control",
    "mutant", "control","mutant", "control", "mutant", "control","mutant",
    "control")
result2 <- logFCsingle(GSE84402, class)

# analysis
listLogFC <- list(result, result2)
listTitle <- c("GSE29721", "GSE84402")
aggreg <- Aggreg(listLogFC, listTitle, padj = 0.01, logFC = 1)

}

}
\references{
Kolde R (2022). \emph{RobustRankAggreg: Methods for Robust Rank Aggregation}.
R package version 1.2.1,
\url{https://CRAN.R-project.org/package=RobustRankAggreg}.

Wang H, Huo X, Yang XR, He J et al. STAT3-mediated upregulation of lncRNA
HOXD-AS1 as a ceRNA facilitates liver cancer metastasis by regulating
SOX4. Mol Cancer 2017 Aug 14;16(1):136. PMID: 28810927

Stefanska B, Huang J, Bhattacharyya B, Suderman M et al. Definition of the
landscape of promoter DNA hypomethylation in liver cancer. Cancer Res
2011 Sep 1;71(17):5891-903. PMID: 21747116
}
